added usuarios permission

+++ resources/views/roles/index.blade.php
@@ -11,9 +11,9 @@
                     <div class="card">
                         <div class="card-body">
         
-                        <!-- @ can('crear-rol') -->
+                        @can('roles-crear')
                         <a class="btn btn-warning" href="{{ route('roles.create') }}">Nuevo</a>                        
-                        <!-- @ endcan -->
+                        @endcan
         
                 
                             <table class="table table-striped mt-2">
@@ -26,15 +26,15 @@
                                 <tr>                           
                                     <td>{{ $role->name }}</td>
                                     <td>                                
-                                        <!-- @ can('editar-rol') -->
+                                        @can('roles-editar') 
                                             <a class="btn btn-primary" href="{{ route('roles.edit',$role->id) }}">Editar</a>
-                                        <!-- @ endcan -->
+                                        @endcan
                                         
-                                        <!-- @ can('borrar-rol') -->
+                                        @can('roles-borrar') 
                                             {!! Form::open(['method' => 'DELETE','route' => ['roles.destroy', $role->id],'style'=>'display:inline']) !!}
                                                 {!! Form::submit('Borrar', ['class' => 'btn btn-danger']) !!}
                                             {!! Form::close() !!}
-                                        <!-- @ endcan -->
+                                        @endcan
                                     </td>
                                 </tr>
                                 @endforeach
